Backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
python-dotenv==0.19.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
joblib==1.1.0 